impl(graph_read): context_graph chain/current/neighbors mode handlers (#608)
Implements all three context_graph traversal modes in mcp/graph_read.rs
(split into graph_read_supersession.rs + graph_read_neighbors.rs per 500-line
limit). SQL CTEs for chain/current (ADR-001), BFS over TypedRelationGraph for
neighbors depth>1 (ADR-005). Adds node_id_for_index accessor to TypedRelationGraph
(ADR-008), read_pool_server() to SqlxStore (C-07), and petgraph dependency.
