+++ crates/unimatrix-server/Cargo.toml
@@ -26,6 +26,8 @@ unimatrix-learn = { path = "../unimatrix-learn" }
 
 http = "1"
 nix = { version = "0.31", features = ["user", "process"] }
+# petgraph: needed for NodeIndex in BFS traversal (neighbors depth>1, ADR-005, vnc-018)
+petgraph = { version = "0.8", default-features = false, features = ["stable_graph"] }
 fs2 = "0.4"
 sqlx = { version = "0.8", features = ["sqlite", "runtime-tokio", "macros"] }
 rmcp = { version = "=0.16.0", features = ["server", "client", "transport-io", "macros"] }
